// src/types/API.ts
export type CreateCommentMutationVariables = {
  condition?: ModelCommentConditionInput | null,
  input: CreateCommentInput,
};

export type CreateCommentMutation = {
  createComment?:  {
    __typename: "Comment",
    comment: string,
    createdAt: string,
    id: string,
    like?:  {
      __typename: "Likes",
      commentId?: string | null,
      createdAt: string,
      id: string,
      owner?: string | null,
      postId?: string | null,
      updatedAt: string,
      userId?: string | null,
    } | null,
    owner?: string | null,
    post?:  {
      __typename: "Post",
      createdAt: string,
      description?: string | null,
      id: string,
      image?: string | null,
      images: Array< string | null >,
      nofComments: number,
      nofLikes: number,
      owner?: string | null,
      updatedAt: string,
      userId?: string | null,
      video?: string | null,
    } | null,
    postId?: string | null,
    updatedAt: string,
    user?:  {
      __typename: "User",
      bio?: string | null,
      createdAt: string,
      email: string,
      id: string,
      image?: string | null,
      name: string,
      nofFollowers: number,
      nofFollowings: number,
      nofPosts: number,
      owner?: string | null,
      updatedAt: string,
      username: string,
      website?: string | null,
    } | null,
    userId?: string | null,
  } | null,
};

export type CreateLikesMutationVariables = {
  condition?: ModelLikesConditionInput | null,
  input: CreateLikesInput,
};

export type CreateLikesMutation = {
  createLikes?:  {
    __typename: "Likes",
    comment?:  {
      __typename: "Comment",
      comment: string,
      createdAt: string,
      id: string,
      owner?: string | null,
      postId?: string | null,
      updatedAt: string,
      userId?: string | null,
    } | null,
    commentId?: string | null,
    createdAt: string,
    id: string,
    owner?: string | null,
    post?:  {
      __typename: "Post",
      createdAt: string,
      description?: string | null,
      id: string,
      image?: string | null,
      images: Array< string | null >,
      nofComments: number,
      nofLikes: number,
      owner?: string | null,
      updatedAt: string,
      userId?: string | null,
      video?: string | null,
    } | null,
    postId?: string | null,
    updatedAt: string,
    user?:  {
      __typename: "User",
      bio?: string | null,
      createdAt: string,
      email: string,
      id: string,
      image?: string | null,
      name: string,
      nofFollowers: number,
      nofFollowings: number,
      nofPosts: number,
      owner?: string | null,
      updatedAt: string,
      username: string,
      website?: string | null,
    } | null,
    userId?: string | null,
  } | null,
};

export type CreatePostMutationVariables = {
  condition?: ModelPostConditionInput | null,
  input: CreatePostInput,
};

export type CreatePostMutation = {
  createPost?:  {
    __typename: "Post",
    comments?:  {
      __typename: "ModelCommentConnection",
      nextToken?: string | null,
    } | null,
    createdAt: string,
    description?: string | null,
    id: string,
    image?: string | null,
    images: Array< string | null >,
    likes?:  {
      __typename: "ModelLikesConnection",
      nextToken?: string | null,
    } | null,
    nofComments: number,
    nofLikes: number,
    owner?: string | null,
    updatedAt: string,
    user?:  {
      __typename: "User",
      bio?: string | null,
      createdAt: string,
      email: string,
      id: string,
      image?: string | null,
      name: string,
      nofFollowers: number,
      nofFollowings: number,
      nofPosts: number,
      owner?: string | null,
      updatedAt: string,
      username: string,
      website?: string | null,
    } | null,
    userId?: string | null,
    video?: string | null,
  } | null,
};

export type CreateUserMutationVariables = {
  condition?: ModelUserConditionInput | null,
  input: CreateUserInput,
};

export type CreateUserMutation = {
  createUser?:  {
    __typename: "User",
    bio?: string | null,
    comments?:  {
      __typename: "ModelCommentConnection",
      nextToken?: string | null,
    } | null,
    createdAt: string,
    email: string,
    id: string,
    image?: string | null,
    likes?:  {
      __typename: "ModelLikesConnection",
      nextToken?: string | null,
    } | null,
    name: string,
    nofFollowers: number,
    nofFollowings: number,
    nofPosts: number,
    owner?: string | null,
    posts?:  {
      __typename: "ModelPostConnection",
      nextToken?: string | null,
    } | null,
    updatedAt: string,
    username: string,
    website?: string | null,
  } | null,
};

export type DeleteCommentMutationVariables = {
  condition?: ModelCommentConditionInput | null,
  input: DeleteCommentInput,
};

export type DeleteCommentMutation = {
  deleteComment?:  {
    __typename: "Comment",
    comment: string,
    createdAt: string,
    id: string,
    like?:  {
      __typename: "Likes",
      commentId?: string | null,
      createdAt: string,
      id: string,
      owner?: string | null,
      postId?: string | null,
      updatedAt: string,
      userId?: string | null,
    } | null,
    owner?: string | null,
    post?:  {
      __typename: "Post",
      createdAt: string,
      description?: string | null,
      id: string,
      image?: string | null,
      images: Array< string | null >,
      nofComments: number,
      nofLikes: number,
      owner?: string | null,
      updatedAt: string,
      userId?: string | null,
      video?: string | null,
    } | null,
    postId?: string | null,
    updatedAt: string,
    user?:  {
      __typename: "User",
      bio?: string | null,
      createdAt: string,
      email: string,
      id: string,
      image?: string | null,
      name: string,
      nofFollowers: number,
      nofFollowings: number,
      nofPosts: number,
      owner?: string | null,
      updatedAt: string,
      username: string,
      website?: string | null,
    } | null,
    userId?: string | null,
  } | null,
};

export type DeleteLikesMutationVariables = {
  condition?: ModelLikesConditionInput | null,
  input: DeleteLikesInput,
};

export type DeleteLikesMutation = {
  deleteLikes?:  {
    __typename: "Likes",
    comment?:  {
      __typename: "Comment",
      comment: string,
      createdAt: string,
      id: string,
      owner?: string | null,
      postId?: string | null,
      updatedAt: string,
      userId?: string | null,
    } | null,
    commentId?: string | null,
    createdAt: string,
    id: string,
    owner?: string | null,
    post?:  {
      __typename: "Post",
      createdAt: string,
      description?: string | null,
      id: string,
      image?: string | null,
      images: Array< string | null >,
      nofComments: number,
      nofLikes: number,
      owner?: string | null,
      updatedAt: string,
      userId?: string | null,
      video?: string | null,
    } | null,
    postId?: string | null,
    updatedAt: string,
    user?:  {
      __typename: "User",
      bio?: string | null,
      createdAt: string,
      email: string,
      id: string,
      image?: string | null,
      name: string,
      nofFollowers: number,
      nofFollowings: number,
      nofPosts: number,
      owner?: string | null,
      updatedAt: string,
      username: string,
      website?: string | null,
    } | null,
    userId?: string | null,
  } | null,
};

export type DeletePostMutationVariables = {
  condition?: ModelPostConditionInput | null,
  input: DeletePostInput,
};

export type DeletePostMutation = {
  deletePost?:  {
    __typename: "Post",
    comments?:  {
      __typename: "ModelCommentConnection",
      nextToken?: string | null,
    } | null,
    createdAt: string,
    description?: string | null,
    id: string,
    image?: string | null,
    images: Array< string | null >,
    likes?:  {
      __typename: "ModelLikesConnection",
      nextToken?: string | null,
    } | null,
    nofComments: number,
    nofLikes: number,
    owner?: string | null,
    updatedAt: string,
    user?:  {
      __typename: "User",
      bio?: string | null,
      createdAt: string,
      email: string,
      id: string,
      image?: string | null,
      name: string,
      nofFollowers: number,
      nofFollowings: number,
      nofPosts: number,
      owner?: string | null,
      updatedAt: string,
      username: string,
      website?: string | null,
    } | null,
    userId?: string | null,
    video?: string | null,
  } | null,
};

export type DeleteUserMutationVariables = {
  condition?: ModelUserConditionInput | null,
  input: DeleteUserInput,
};

export type DeleteUserMutation = {
  deleteUser?:  {
    __typename: "User",
    bio?: string | null,
    comments?:  {
      __typename: "ModelCommentConnection",
      nextToken?: string | null,
    } | null,
    createdAt: string,
    email: string,
    id: string,
    image?: string | null,
    likes?:  {
      __typename: "ModelLikesConnection",
      nextToken?: string | null,
    } | null,
    name: string,
    nofFollowers: number,
    nofFollowings: number,
    nofPosts: number,
    owner?: string | null,
    posts?:  {
      __typename: "ModelPostConnection",
      nextToken?: string | null,
    } | null,
    updatedAt: string,
    username: string,
    website?: string | null,
  } | null,
};

export type UpdateCommentMutationVariables = {
  condition?: ModelCommentConditionInput | null,
  input: UpdateCommentInput,
};

export type UpdateCommentMutation = {
  updateComment?:  {
    __typename: "Comment",
    comment: string,
    createdAt: string,
    id: string,
    like?:  {
      __typename: "Likes",
      commentId?: string | null,
      createdAt: string,
      id: string,
      owner?: string | null,
      postId?: string | null,
      updatedAt: string,
      userId?: string | null,
    } | null,
    owner?: string | null,
    post?:  {
      __typename: "Post",
      createdAt: string,
      description?: string | null,
      id: string,
      image?: string | null,
      images: Array< string | null >,
      nofComments: number,
      nofLikes: number,
      owner?: string | null,
      updatedAt: string,
      userId?: string | null,
      video?: string | null,
    } | null,
    postId?: string | null,
    updatedAt: string,
    user?:  {
      __typename: "User",
      bio?: string | null,
      createdAt: string,
      email: string,
      id: string,
      image?: string | null,
      name: string,
      nofFollowers: number,
      nofFollowings: number,
      nofPosts: number,
      owner?: string | null,
      updatedAt: string,
      username: string,
      website?: string | null,
    } | null,
    userId?: string | null,
  } | null,
};

export type UpdateLikesMutationVariables = {
  condition?: ModelLikesConditionInput | null,
  input: UpdateLikesInput,
};
// ...
export type UpdatePostMutationVariables = {
  condition?: ModelPostConditionInput | null,
  input: UpdatePostInput,
};

export type UpdatePostMutation = {
  updatePost?:  {
    __typename: "Post",
    comments?:  {
      __typename: "ModelCommentConnection",
      nextToken?: string | null,
    } | null,
    createdAt: string,
    description?: string | null,
    id: string,
    image?: string | null,
    images: Array< string | null >,
    likes?:  {
      __typename: "ModelLikesConnection",
      nextToken?: string | null,
    } | null,
    nofComments: number,
    nofLikes: number,
    owner?: string | null,
    updatedAt: string,
    user?:  {
      __typename: "User",
      bio?: string | null,
      createdAt: string,
      email: string,
      id: string,
      image?: string | null,
      name: string,
      nofFollowers: number,
      nofFollowings: number,
      nofPosts: number,
      owner?: string | null,
      updatedAt: string,
      username: string,
      website?: string | null,
    } | null,
    userId?: string | null,
    video?: string | null,
  } | null,
};